+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] -2019-05-16
+### Added
+- Add return button in settings page
+
+### Change
+- Shorten uploaded file name
+
+### Fix
+- Fix image scaling when image changed
+
 ## [0.2.0] -2019-05-15
 ### Fix
 - Fix production in gitlab pages